Crate libra_logger

A default logger for Libra project.

Usage

use libra_logger::prelude::*;

pub fn main() {
    let _g = libra_logger::set_default_global_logger(false /* async */, Some(256));
    info!("Starting...");
}

Modules

prelude
Logger prelude which includes all logging macros.

Macros

crit
Log a critical level message using current scope logger

debug
Log a debug level message using current scope logger

error
Log a error level message using current scope logger

info
Log a info level message using current scope logger

slog_crit
Log critical level record (alias)

slog_debug
Log debug level record (alias)

slog_error
Log error level record

slog_info
Log info level record (alias)

slog_trace
Log trace level record (alias)

slog_warn
Log warning level record (alias)

trace
Log a trace level message using current scope logger

warn
Log a warning level message using current scope logger

Functions

init_for_e2e_testing
Create and setup default global logger for use in end to end testing.

set_default_global_logger
Creates and sets default global logger. Caller must keep the returned guard alive.

set_simple_logger
Simple logger mostly intended for use in test code It contains bunch of boilerplate hacks to make output of test look not too verbose(but still have some logs)

set_simple_logger_prefix
Sets thread local prefix for log lines This is useful for tests that simulates multiple machines Logger prefix can be used to separate logs from those multiple simulated instances

try_init_for_testing
Create and setup default global logger following the env-logger conventions, i.e. configured by environment variable RUST_LOG. This is useful to make logging optional in unit tests.