Crate libra_logger

This crates provides API for both structured and not structured(text) logging

Text logging

Text logging is configured via RUST_LOG macro and have exactly same facade as rust log crate

Structured logging

This crate contains two levels of API for structured logging

StructuredLogEntry class and send_struct_log! macro for directly composing structured log

Bridge between traditional log! macro and structured logging API

Configuration

Structured logger has separate log levels, configured with STRUCT_LOG_LEVEL. It is set to debug by default, but it only has effect if structured logger is initialized.

Structured logger can be initialized manually with one of init_XXX_struct_log functions. Preferred way to initialize structured logging is by using init_struct_log_from_env. In this case STRUCT_LOG_FILE environment variable is used to set file name for structured logs

Direct API

use std::collections::HashMap;
use serde_json::Value;

pub struct StructuredLogEntry {
    text: Option<String>,
    pattern: Option<&'static str>,
    name: Option<&'static str>,
    module: Option<&'static str>,
    location: Option<&'static str>,
    git_rev: Option<&'static str>,
    data: HashMap<&'static str, Value>,
}

impl StructuredLogEntry {
    pub fn new_unnamed() -> Self { /* ... */ }
    pub fn new_named(name: &'static str) -> Self { /* ... */ }
    /* ..Bunch of builder style setters for chained initialization such as
        entry.data(a, b).data(x, y).. */
}

// Usage:
send_struct_log!(StructuredLogEntry::new_named("Committed")
   .data("block", &block)
   .data("autor", &author))

Arguments passed to .data will be serialized into json, and as such should implement Serialize. Only static strings are allowed as field names.

send_struct_log! should be used to send structured log entry. This macro populates metadata such as git_rev, location and module, and also skips evaluation of StructuredLogEntry entirely, if structured logging is disabled

Log macro bridge

Crate owners are not required to rewrite their code right away to support new structured logging. Importing logger crate will automatically emit structured logging on every log(debug!, info!, etc) macro invocation, unless no_struct_log feature was enabled on this crate.

Note: Only enable 'no_struct_log' on leaf binary crates, never on library, as it might disable structured logging for other crates, due to how cargo handles features at this moment.

So

info!("Committing {}", block);
// Will emit(in addition to regular text log) structured log such as
// {
//   pattern: "Committing {}",
//   data: {
//     block: "<id>"
//   },
//   ...metadata...
// }

There are few caveats to automatic structured logging generation

Argument values for structured logging will be serialized with Debug implementation(vs proper json serialization if send_struct_log! macro is used), regardless of what formatter is used for textual log. As a consequence, this means, that every log argument must implement Debug. In theory, this is extra limitation, but in practice currently in libra node crate and it's dependencies there was no single log argument that would not satisfy this criteria.

Field names will be automatically evaluated if expression is a single identifier, as in example above field block will get human readable name in json. However, if more complex expression is passed, structured log field name will be based on position of argument: "_0", "_1", etc:

info!("Committing {}", block.id());
//->
// {
//  data: {
//    "_0": "<id>"
//  },
//  ...metadata...
// }

Another way to set proper name to json fields is to use named format arguments:

info!("Committing {id}", id=block.id());
//->
// {
//  data: {
//    "id": "<id>"
//  },
//  ...metadata...
// }

Structured log sink Application must define implementation of StructLogSink interface in order to direct structured logs emitted by send_struct_log and other macros. This sink can be only initialized once, by calling set_struct_logger function.

Currently 3 implementations for StructLogSink exist:

NopStructLog ignores structured logs

PrintStructLog immediately prints structured logs to stdout

FileStructLog prints structured logs into provided file. Using this logger creates separate thread for writing files and structured logging itself is asynchronous in this case.

Re-exports

pub use log;

Modules

prelude

Macros

crit
Define crit macro that specify libra as the target

debug
Define debug macro that specify libra as the target

error
Define macro that specify libra as the target

format_index

format_struct_arg

format_struct_args

format_struct_args_and_pattern

git_rev

info
Define info macro that specify libra as the target

location

send_struct_log

struct_log

struct_log_enabled

trace
Define trace macro that specify libra as the target

warn
Define warn macro that specify libra as the target

Structs

Logger
Logging framework for Libra that encapsulates a minimal dependency logger with support for environmental variable (RUST_LOG) and asynchronous logging. Note: only a single logger can be instantiated at a time. Repeated instantiates of the loggers will only adjust the global logging level but will not change the initial filter.

StructuredLogEntry

Enums

Level
An enum representing the available verbosity levels of the logger.

Constants

CHANNEL_SIZE

DEFAULT_TARGET

Traits

StructLogSink

Functions

init_file_struct_log
Initializes struct logger sink that writes to specified file. Can only be called once

init_println_struct_log
Initialize struct logger sink that prints all structured logs to stdout Can only be called once

init_struct_log_from_env
Initializes struct logger from STRUCT_LOG_FILE env var. If STRUCT_LOG_FILE is set, STRUCT_LOG_UDP_ADDR will be ignored. Can only be called once

set_struct_logger
Sets structured logger

struct_logger_enabled
Checks if structured logging is enabled for level

struct_logger_set
Checks if structured logging is enabled