Crate libra_logger

This crates provides an API for both structured and non-structured(text) logging.

Text logging

Text logging is configured via RUST_LOG environment variable and has exactly same facade as the rust log crate.

Structured logging

This crate contains two types of APIs for structured logging.

The StructuredLogEntry class and (sl_info!, sl_error!, ...) macros for directly composing structured logs.

A bridge between traditional log! macro and the structured logging API (which will be deprecated).

Configuration

Structured logger has separate log levels that are configured with STRUCT_LOG_LEVEL. It is set to DEBUG by default, but executes if structured logger is initialized.

Structured logger can be initialized manually with one of the init_XXX_struct_log() functions. The preferred way to initialize structured logging is by using init_struct_log_from_env(). In this case, the STRUCT_LOG_FILE environment variable is used to set the file name for structured logs.

Direct API

use std::collections::HashMap;
use serde_json::Value;

pub struct StructuredLogEntry {
    id: String,
    log: Option<String>,
    pattern: Option<&'static str>,
    category: Option<&'static str>,
    name: Option<&'static str>,
    module: Option<&'static str>,
    location: Option<&'static str>,
    timestamp: String,
    level: Option<log::Level>,
    data: HashMap<&'static str, Value>,
}

impl StructuredLogEntry {
    pub fn new_named(category: &'static str, name: &'static str) -> Self { /* ... */ }
    /* ... Builder style setters for chained initialization such as
        entry.data(a, b).data(x, y) ... */
}

// Usage:
let logging_field: LoggingField<String> = LoggingField::new("String");
let string = "test".to_string();

sl_info!(StructuredLogEntry::new_named("TransactionEvents", "Committed")
   .data("block", &block)
   .data_display("author", &author)
   .field(&logging_field, &string)
   .log(format!("Committed block: {:?} by {}", block, author))

Arguments passed to data() will be serialized into JSON, and must implement Serialize. Arguments passed to data_display() will instead use display and must implement Display. Arguments passed to field() allows for typed fields for type checking, and must implement Serialize. Only static strings are allowed as field names.

(sl_info!, sl_error!, etc.) should be used to send structured log entries based on log level. This macro populates the metadata such as code location and module, and skips the evaluation of StructuredLogEntry entirely if structured logging is disabled for the log level.

Typed Schema's

The Schema trait can be used to implement typed logging schemas. This can either be implemented by hand or derived using the Schema derive proc-macro, implementing the Schema trait for the struct as well as providing setters for all fields.

use libra_logger::Schema;

#[derive(Schema)]
struct LogSchema<'a> {
    // Log using this type's Serialize impl
    a: usize,
    // Log using this type's Debug impl
    #[schema(debug)]
    b: Option<Vec<bool>>,
    // Log using this type's Display impl
    #[schema(display)]
    c: Option<&'a str>,
}

Log macro bridge

Crate owners are not required to rewrite their code right away to support new structured logging. Importing logger crate will automatically emit structured logs on every log(debug!, info!, etc.) macro invocation.

So

info!("Committing {}", block);
// Will emit(in addition to regular text log) structured log such as
// {
//   level: "Info",
//   pattern: "Committing {}",
//   data: {
//     block: "<id>"
//   },
//   ...metadata...
// }

There are few caveats to automatic structured logging generation:

Argument values for structured logging will be serialized using Debug vs Serialize, regardless of what formatter is used for the text log. As a consequence, every log argument must implement Debug. This has led to unexpected large logs from the default Debug implementations.

Field names will be automatically evaluated if the expression is a single identifier, as in the example above, the field block will be named block. However, if a more complex expression is passed (e.g. block.id()), the field name will be based on the position of the argument: _0, _1, etc.

info!("Committing {}", block.id());
//->
// {
//  data: {
//    "_0": "<id>"
//  },
//  ...metadata...
// }

Another way to set the name for fields is to use named format arguments:

info!("Committing {id}", id=block.id());
//->
// {
//  data: {
//    "id": "<id>"
//  },
//  ...metadata...
// }

Structured Log Sink

The application must define implementation of the StructLogSink trait in order to direct structured logs emitted by sl_info! and other macros. The global sink can be only initialized once, by calling the set_struct_logger() function.

Currently 4 implementations for StructLogSink exist:

NopStructLog ignores structured logs

PrintStructLog immediately prints structured logs to stdout

FileStructLog prints structured logs into a file. This logger separate thread for writing files asynchronously.

TCPStructLog sends structured logs to a TCP endpoint. This logger separate thread for sending logs asynchronously.

Re-exports

pub use log;

Modules

counters

json_log

prelude

Macros

debug
Define debug macro that specify libra as the target

error
Define error macro that specify libra as the target

event
Writes event to event stream Example: event!("committed", block="b");

format_index

format_struct_arg

format_struct_args

format_struct_args_and_pattern

info
Define info macro that specify libra as the target

location

sl_debug

sl_error

sl_info

sl_level
Allows for dynamic macro levels, but still filtering Use of this is highly discouraged, and you should stick to the sl_info! type macros.

sl_trace

sl_warn

struct_log_enabled

text_to_struct_log
Converts text logs to unnamed / uncategorized structured logs

trace
Define trace macro that specify libra as the target

warn
Define warn macro that specify libra as the target

Structs

Key

Logger
Logging framework for Libra that encapsulates a minimal dependency logger with support for environmental variable (RUST_LOG) and asynchronous logging. Note: only a single logger can be instantiated at a time. Repeated instantiates of the loggers will only adjust the global logging level but will not change the initial filter.

LoggingField
Field is similar to .data but restricts type of the value to a specific type.

StructuredLogEntry

Enums

Level
An enum representing the available verbosity levels of the logger.

Value

Constants

CHANNEL_SIZE

DEFAULT_TARGET

Traits

Schema
A schema of key-value pairs.

StructLogSink

Visitor
A visitor for the key-value pairs in a Schema.

Functions

init_file_struct_log
Initializes struct logger sink that writes to specified file. Can only be called once

init_println_struct_log
Initialize struct logger sink that prints all structured logs to stdout Can only be called once

init_struct_log_from_env
Initializes struct logger from STRUCT_LOG_FILE env var. If STRUCT_LOG_FILE is set, STRUCT_LOG_TCP_ADDR will be ignored. Can only be called once

set_struct_logger
Sets structured logger

struct_logger_enabled
Checks if structured logging is enabled for level

struct_logger_set
Checks if structured logging is enabled

Derive Macros

Schema